fix: null handling and more revamps

+++ main.h
@@ -4,7 +4,7 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-int print_int(int intv);
+int print_int(long int intv);
 int _puts(char *s);
